src/requirements.txt
# requirements.txt - Main Python Dependencies
kafka-python==2.0.2
pyspark==3.4.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.11.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1
flower==2.0.1
python-dotenv==1.0.0
faker==19.3.1
pydantic==2.1.1
fastapi==0.101.1
uvicorn==0.23.2
loguru==0.7.0